move instance_norm from fluid to static
