+++ python/paddle/fluid/layers/nn.py
@@ -73,7 +73,6 @@ __all__ = [
     'pool2d',
     'pool3d',
     'batch_norm',
-    'instance_norm',
     'data_norm',
     'reduce_mean',
     'reduce_all',
@@ -2473,145 +2472,6 @@ def batch_norm(
     )
 
     return helper.append_activation(batch_norm_out)
-
-
-def instance_norm(
-    input, epsilon=1e-05, param_attr=None, bias_attr=None, name=None
-):
-    r"""
-    :api_attr: Static Graph
-
-    **Instance Normalization Layer**
-
-    Can be used as a normalizer function for convolution or fully_connected operations.
-    The required data format for this layer is one of the following:
-
-    DataLayout: NCHW `[batch, in_channels, in_height, in_width]`
-
-    Refer to `Instance Normalization: The Missing Ingredient for
-    Fast Stylization <https://arxiv.org/pdf/1607.08022.pdf>`_
-    for more details.
-
-    :math:`input` is the input features over a mini-batch.
-
-    ..  math::
-
-        \\mu_{\\beta} &\\gets \\frac{1}{HW} \\sum_{i=1}^{HW} x_i \\qquad &//\\
-        \\ mean\ of\ one\  feature\ map\ in\ mini-batch \\\\
-        \\sigma_{\\beta}^{2} &\\gets \\frac{1}{HW} \\sum_{i=1}^{HW}(x_i - \\
-        \\mu_{\\beta})^2 \\qquad &//\ variance\ of\ one\ feature\ map\ in\ mini-batch \\\\
-        \\hat{x_i} &\\gets \\frac{x_i - \\mu_\\beta} {\\sqrt{\\
-        \\sigma_{\\beta}^{2} + \\epsilon}} \\qquad &//\ normalize \\\\
-        y_i &\\gets \\gamma \\hat{x_i} + \\beta \\qquad &//\ scale\ and\ shift
-
-    Note:
-        `H` means height of feature map, `W` means width of feature map.
-
-    Args:
-        input(Tensor): The rank of input tensor can be 2, 3, 4, 5.
-            The data type is float32 or float64.
-        epsilon(float, Default 1e-05): A value added to the denominator for
-            numerical stability. Default is 1e-5.
-        param_attr(ParamAttr|None|bool, optional): The parameter attribute for Parameter `scale`
-             of instance_norm. If it is set to None or one attribute of ParamAttr, instance_norm
-	     will create ParamAttr as param_attr, the name of scale can be set in ParamAttr.
-	     If the Initializer of the param_attr is not set, the parameter is initialized
-	     with Xavier. If the param_attr is set to False, instance_norm will not create param_attr.
-             Default: None.
-        bias_attr(ParamAttr|None|bool, optional): The parameter attribute for the bias of instance_norm.
-             If it is set to None or one attribute of ParamAttr, instance_norm
-	     will create ParamAttr as bias_attr, the name of bias can be set in ParamAttr.
-	     If the Initializer of the bias_attr is not set, the bias is initialized zero.
-             If the bias_attr is set to False, instance_norm will not create bias_attr.
-	     Default: None.
-        name(string, Default None): A name for this layer(optional). If set None, the layer
-            will be named automatically.
-
-    Returns:
-        A Tensor which is the result after applying instance normalization on the input,
-        has same shape and data type with input.
-
-    Examples:
-
-        .. code-block:: python
-
-            import paddle
-            paddle.enable_static()
-            x = paddle.static.data(name='x', shape=[3, 7, 3, 7], dtype='float32')
-            hidden1 = paddle.static.nn.fc(x, size=200)
-            hidden2 = paddle.static.nn.instance_norm(hidden1)
-    """
-    check_variable_and_dtype(
-        input, 'input', ['float32', 'float64'], 'instance_norm'
-    )
-    if param_attr is False:
-        assert (
-            bias_attr is False
-        ), "param_attr and bias_attr must be set to False at the same time in instance_norm"
-
-    helper = LayerHelper('instance_norm', **locals())
-    dtype = helper.input_dtype()
-
-    # use fp32 for in parameter
-    if dtype == core.VarDesc.VarType.FP16:
-        dtype = core.VarDesc.VarType.FP32
-
-    input_shape = input.shape
-    if len(input.shape) < 2 or len(input.shape) > 5:
-        raise ValueError(
-            'expected 2D or 3D or 4D or 5D input (got {}D input, input shape is: {})'.format(
-                len(input.shape), input_shape
-            )
-        )
-    channel_num = input_shape[1]
-
-    param_shape = [channel_num]
-
-    if param_attr != False and bias_attr != False:
-        # create parameter
-        scale = helper.create_parameter(
-            attr=helper.param_attr,
-            shape=param_shape,
-            dtype=dtype,
-            default_initializer=Constant(1.0),
-        )
-        bias = helper.create_parameter(
-            attr=helper.bias_attr,
-            shape=param_shape,
-            dtype=dtype,
-            is_bias=True,
-            default_initializer=Constant(0.0),
-        )
-
-    # create output
-    saved_mean = helper.create_variable_for_type_inference(
-        dtype=dtype, stop_gradient=True
-    )
-    saved_variance = helper.create_variable_for_type_inference(
-        dtype=dtype, stop_gradient=True
-    )
-
-    instance_norm_out = helper.create_variable_for_type_inference(dtype)
-
-    inputs = {"X": input}
-    if param_attr != False and bias_attr != False:
-        inputs["Scale"] = scale
-        inputs["Bias"] = bias
-
-    helper.append_op(
-        type="instance_norm",
-        inputs=inputs,
-        outputs={
-            "Y": instance_norm_out,
-            "SavedMean": saved_mean,
-            "SavedVariance": saved_variance,
-        },
-        attrs={
-            "epsilon": epsilon,
-        },
-    )
-
-    return instance_norm_out
 
 
 @static_only
