+++ python/paddle/fluid/tests/unittests/ir/inference/test_trt_instance_norm_op.py
@@ -20,6 +20,7 @@ import unittest
 import numpy as np
 from inference_pass_test import InferencePassTest
 
+import paddle
 import paddle.fluid as fluid
 import paddle.fluid.core as core
 from paddle.fluid.core import AnalysisConfig, PassVersionChecker
@@ -43,7 +44,7 @@ class TRTInstanceNormTest(InferencePassTest):
         with fluid.program_guard(self.main_program, self.startup_program):
             shape = [-1, self.channel, self.height, self.width]
             data = fluid.data(name='in', shape=shape, dtype='float32')
-            instance_norm_out = fluid.layers.instance_norm(data)
+            instance_norm_out = paddle.static.nn.instance_norm(data)
             out = fluid.layers.batch_norm(instance_norm_out, is_test=True)
 
         shape[0] = self.bs
