+++ python/paddle/fluid/tests/unittests/test_imperative_star_gan_with_gradient_penalty.py
@@ -120,7 +120,7 @@ class InstanceNorm(fluid.dygraph.Layer):
             )
             return out
         else:
-            return fluid.layers.instance_norm(
+            return paddle.static.nn.instance_norm(
                 input,
                 epsilon=self.epsilon,
                 param_attr=fluid.ParamAttr(self.scale.name),
